requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
elevenlabs==0.2.26
gunicorn==21.2.0
python-multipart==0.0.6
mistralai==0.0.12 